// condor/client/ng/cases/case-view/view-data.model.ts
export class CaseViewViewData {
    model?: any;
    caseId?: number;
    caseKey?: number;
    irn?: string;
    title?: string;
    propertyType?: string;
    country?: string;
    caseCategory?: string;
    subType?: string;
    basis?: string;
    status?: string;
    caseStatus?: string;
    renewalStatus?: string;
    officialNumber?: string;
    family?: string;
    caseType?: string;
    caseOffice?: string;
    fileLocation?: string;
    profitCentre?: string;
    localClientFlag?: boolean;
    hasBillPercentageDisplayed?: Array<string>;
    entitySize?: string;
    typeOfMark?: string;
    numberInSeries?: number;
    staff?: any;
    classes?: string;
    firstApplicant?: string;
    instructor?: string;
    applicationFilingDate?: any;
    caseImageData?: any;
    imageKey?: number;
    imageDescription?: string;
    isRegistered?: boolean;
    isPending?: boolean;
    isDead?: boolean;
    propertyTypeCode?: string;
    propertyTypeImageId?: number;
    policingStatus?: string;
    displayNameVariants?: boolean;
    yourReference?: string;
    caseDefaultDescription?: string;
    clientMainContact?: any;
    ourContact?: any;
    allowSubClass?: boolean;
    useDefaultCountryForClasses?: boolean;
    allowSubClassWithoutItem?: boolean;
    hasAccessToAttachmentSubject?: boolean;
    hasCaseEventAuditingConfigured?: boolean;
    canMaintainCaseEvent?: boolean;
    maintainEventNotes?: boolean;
    clearCaseEventDates?: boolean;
    canViewRuleDetails?: boolean;
    canAddAttachment?: boolean;
}